Move the main function for config building to a separate module and add a script to run a batch of jobs on local resources
